Fix reference to props in constructor
In <=IE10, `this.props` is null, so when attempting to set the namespaces,
the reference to `this.props.ns` raises an exception. By updating the
reference to use the version passed into the constructor, we are
consistent with other usages of props on the constructor, and I've
confirmed this patch works in <=IE10 as well.