feat: use const type parameters for useTranslation()
TypeScript 5.0 has introduced const type parameters [1]. These indicate
that a more specific type is preferred over a more general one. As
a result, the resulting UseTranslationResponse will be more precise
and retain the passed-in namespace and prefix. This can be used by
extractors and other tooling to find out exactly which key is
accessed by a translation function call.

[1]: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-0.html#const-type-parameters